Adapt the tests to the 3.17 refacto
